updates tooling updates tests updates codestyle address laminas migration
