+++ .docheader
@@ -1,0 +1,3 @@
+@copyright Copyright (c) 2016 bushbaby multimedia. (https://bushbaby.nl)
+@author    Bas Kamer <baskamer@gmail.com>
+@license   MIT
