+++ .gitignore
@@ -3,3 +3,6 @@ composer.phar
 composer.lock
 phpunit.xml
 vendor/
+/.phpunit.result.cache
+/.php_cs.cache
+/build/
