+++ .travis.yml
@@ -1,29 +1,42 @@
-sudo: false
-
 language: php
 
+cache:
+  directories:
+    - $HOME/.composer/cache
+    - $HOME/.php-cs-fixer
+
 matrix:
-  fast_finish: true
   include:
-    - php: 5.5
-    - php: 5.6
-    - php: 7
-    - php: hhvm
-  allow_failures:
-    - php: hhvm
+    - php: 7.2
+      env:
+        - DEPENDENCIES=""
+        - EXECUTE_CS_CHECK=true
+        - TEST_COVERAGE=true
+    - php: 7.2
+      env:
+        - DEPENDENCIES="--prefer-lowest --prefer-stable"
+    - php: 7.3
+      env:
+        - DEPENDENCIES=""
+    - php: 7.3
+      env:
+        - DEPENDENCIES="--prefer-lowest --prefer-stable"
+    - php: 7.4
+      env:
+        - DEPENDENCIES=""
+    - php: 7.4
+      env:
+        - DEPENDENCIES="--prefer-lowest --prefer-stable"
 
 before_script:
+  - mkdir -p "$HOME/.php-cs-fixer"
   - composer self-update
-  - composer install --prefer-source
+  - composer update $DEPENDENCIES
 
 script:
-  - vendor/bin/phpunit -c ./phpunit.xml
-  - vendor/bin/phpcs --standard=PSR2 ./src/
+  - mkdir -p build/logs
+  - if [[ $TEST_COVERAGE == 'true' ]]; then php -dzend_extension=xdebug.so ./vendor/bin/phpunit --coverage-text --coverage-clover ./build/logs/clover.xml; else ./vendor/bin/phpunit; fi
+  - if [[ $EXECUTE_CS_CHECK == 'true' ]]; then ./vendor/bin/php-cs-fixer fix -v --diff --dry-run; fi
 
-after_script:
-  - wget https://scrutinizer-ci.com/ocular.phar
-  - php ocular.phar code-coverage:upload --format=php-clover test/_build/logs/clover.xml
-
-notifications:
-  irc: "irc.freenode.org#zftalk.modules"
-  email: false
+after_success:
+  - if [[ $TEST_COVERAGE == 'true' ]]; then travis_retry php vendor/bin/php-coveralls -v; fi
