+++ composer.json
@@ -1,30 +1,40 @@
 {
-    "name": "bushbaby/doctrine-managerregistry-servicemanager",
-    "description": "An implementation of Doctrine's ManagerRegistry for the ZendFramework ServiceManager",
-    "require": {
-        "php": "^5.5 || ^7.0",
-        "zendframework/zend-servicemanager" : "^2.6 || ^3.0",
-        "doctrine/orm": "^2.4"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^5.1",
-        "squizlabs/php_codesniffer": "^2.5"
-    },
-    "autoload": {
-        "psr-4": {
-            "BsbDoctrineRegistry\\": "src/"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "BsbDoctrineRegistryTest\\": "test/src/"
-        }
-    },
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "Bas Kamer",
-            "email": "bas@bushbaby.nl"
-        }
-    ]
+  "name": "bushbaby/doctrine-managerregistry-servicemanager",
+  "description": "An implementation of Doctrine's ManagerRegistry for the Laminas ServiceManager",
+  "require": {
+    "php": "^7.2",
+    "laminas/laminas-servicemanager": "^3.0",
+    "doctrine/orm": "^2.4"
+  },
+  "require-dev": {
+    "bushbaby/php-cs-fixer-config": "^1.0",
+    "phpunit/phpunit": "^8.5",
+    "php-coveralls/php-coveralls": "^2.0"
+  },
+  "autoload": {
+    "psr-4": {
+      "BsbDoctrineRegistry\\": "src/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "BsbDoctrineRegistryTest\\": "test/"
+    }
+  },
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Bas Kamer",
+      "email": "baskamer@gmail.com"
+    }
+  ],
+  "scripts": {
+    "check": [
+      "@cs",
+      "@test"
+    ],
+    "cs": "php-cs-fixer fix -v --diff --dry-run",
+    "cs-fix": "php-cs-fixer fix -v --diff",
+    "test": "phpunit"
+  }
 }
