+++ test/Registry/ManagerRegistryTest.php
@@ -1,0 +1,21 @@
+<?php
+
+/**
+ * @copyright Copyright (c) 2016 bushbaby multimedia. (https://bushbaby.nl)
+ * @author    Bas Kamer <baskamer@gmail.com>
+ * @license   MIT
+ */
+
+declare(strict_types=1);
+
+namespace BsbDoctrineRegistryTest\Registry;
+
+use PHPUnit\Framework\TestCase;
+
+class ManagerRegistryTest extends TestCase
+{
+    public function testHelp()
+    {
+        // help..! You know how to test this? please feel free. :-)
+    }
+}
