--- test/src/Registry/ManagerRegistryTest.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace BsbDoctrineRegistryTest\Registry;
-
-use BsbDoctrineRegistry\Container\ManagerRegistryFactory;
-use Interop\Container\ContainerInterface;
-use Zend\Diactoros\Response;
-use Zend\ServiceManager\ServiceLocatorInterface;
-
-class ManagerRegistryTest extends \PHPUnit_Framework_TestCase
-{
-
-    public function testHelp()
-    {
-        // help..! You know how to test this? please feel free. :-)
-    }
-}
